requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
python-dotenv>=1.0.0
langchain>=0.0.267
langchain-openai>=0.0.2
langchain-community>=0.0.10
langchain-chroma>=0.0.10
chromadb>=0.4.13
python-multipart>=0.0.6
pydantic>=2.4.2
pypdf>=3.15.1
docx2txt>=0.8